re located folders to help next refactoring phase which will require more folders/files
