+++ app/javascript/components/About/JimmysTips/JimmysTipsContainer.jsx
