+++ app/javascript/components/About/Portfolio/PortfolioContainer.jsx
@@ -3,7 +3,6 @@ import PortfolioProperties from "./PortfolioProperties";
 import {
   FadeIn,
   ParallaxBannerRoutes,
-  FormMaps,
   ParallaxEditForm
 } from "../../Constants/Constants";
