+++ app/javascript/components/App.jsx
@@ -6,30 +6,30 @@ import {
   Redirect
 } from "react-router-dom";
 import Home from "../components/Home/Home";
-import Registration from "../components/Home/User/Registration";
-import Login from "../components/Home/User/Login";
-import AboutContainer from "../components/Home/About/AboutContainer";
-import AboutCompanyContainer from "../components/Home/AboutCompany/AboutCompanyContainer";
-import JimmysTipsContainer from "../components/Home/JimmysTips/JimmysTipsContainer";
-import ContactContainer from "../components/Home/Contact/ContactContainer";
-import NavbarContainer from "../components/Home/Navbar/NavbarContainer";
-import NewTown from "../components/Home/Town/NewTown";
+import Registration from "./User/Registration";
+import Login from "./User/Login";
+import AboutContainer from "./About/AboutJimmy/AboutContainer";
+import AboutCompanyContainer from "./About/AboutCompany/AboutCompanyContainer";
+import JimmysTipsContainer from "./About/JimmysTips/JimmysTipsContainer";
+import ContactContainer from "./Contact/ContactContainer";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { fab } from "@fortawesome/free-brands-svg-icons";
 import { faTrashAlt, faEdit } from "@fortawesome/free-solid-svg-icons";
-import FooterContainer from "../components/Home/Footer/FooterContainer";
-import TownShowPage from "../components/Home/Town/TownShowPage";
-import EditTown from "../components/Home/Town/EditTown";
-import PortfolioContainer from "./Home/Portfolio/PortfolioContainer";
-import HomeWorthContainer from "../components/Home/Service/HomeWorthContainer";
-import RelocationContainer from "../components/Home/Service/RelocationContainer";
-import BuyingHomeContainer from "../components/Home/Service/BuyingHomeContainer";
-import SellingHomeContainer from "../components/Home/Service/SellingHomeContainer";
-import MarketReportsContainer from "../components/Home/Service/MarketReportsContainer";
-import TestimonialsContainer from "./Home/Testimonials/TestimonialsContainer";
-import NewPartner from "../components/Home/JimmyPartners/NewPartner";
-import PartnerShowPage from "../components/Home/JimmyPartners/PartnerShowPage";
-import EditPartner from "../components/Home/JimmyPartners/EditPartner";
+import NavbarContainer from "./Navbar/NavbarContainer";
+import FooterContainer from "./Footer/FooterContainer";
+import NewTown from "./FeaturedCommunities/Town/NewTown";
+import TownShowPage from "./FeaturedCommunities/Town/TownShowPage";
+import EditTown from "./FeaturedCommunities/Town/EditTown";
+import PortfolioContainer from "./About/Portfolio/PortfolioContainer";
+import HomeWorthContainer from "./Service/HomeWorth/HomeWorthContainer";
+import RelocationContainer from "./Service/Relocation/RelocationContainer";
+import BuyingHomeContainer from "./Service/BuyingHome/BuyingHomeContainer";
+import SellingHomeContainer from "./Service/SellingHome/SellingHomeContainer";
+import MarketReportsContainer from "./Service/MarketReports/MarketReportsContainer";
+import TestimonialsContainer from "./About/Testimonials/TestimonialsContainer";
+import NewPartner from "./JimmyPartners/NewPartner";
+import PartnerShowPage from "./JimmyPartners/PartnerShowPage";
+import EditPartner from "./JimmyPartners/EditPartner";
 
 library.add(fab, faTrashAlt, faEdit);
