+++ app/javascript/components/JimmyPartners/EditPartner.jsx
