+++ app/javascript/components/JimmyPartners/PartnerList.jsx
