+++ app/javascript/components/Service/MarketReports/MarketPhotoContainer.jsx
