+++ app/javascript/components/Service/Relocation/RelocationPhotoContainer.jsx
