+++ app/javascript/components/User/Registration.jsx
